# pkg/api/listings/listings.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Marketplace Listings API",
    "description": "API для управления и просмотра объявлений маркетплейса",
    "version": "1.0.0"
  },
  "tags": [
    {
      "name": "ListingsService"
    }
  ],
  "host": "localhost:8080",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/listings": {
      "get": {
        "summary": "Получение ленты объявлений",
        "description": "Возвращает ленту объявлений с возможностью сортировки, фильтрации и пагинации",
        "operationId": "ListingsService_GetListings",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/listingsListingsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "page",
            "description": "Пагинация",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "perPage",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "sortBy",
            "description": "Сортировка",
            "in": "query",
            "required": false,
            "type": "string",
            "enum": [
              "SORT_FIELD_UNSPECIFIED",
              "SORT_FIELD_CREATED_AT",
              "SORT_FIELD_PRICE"
            ],
            "default": "SORT_FIELD_UNSPECIFIED"
          },
          {
            "name": "sortOrder",
            "in": "query",
            "required": false,
            "type": "string",
            "enum": [
              "SORT_ORDER_UNSPECIFIED",
              "SORT_ORDER_ASC",
              "SORT_ORDER_DESC"
            ],
            "default": "SORT_ORDER_UNSPECIFIED"
          },
          {
            "name": "minPrice",
            "description": "Фильтрация по цене",
            "in": "query",
            "required": false,
            "type": "number",
            "format": "float"
          },
          {
            "name": "maxPrice",
            "in": "query",
            "required": false,
            "type": "number",
            "format": "float"
          }
        ],
        "tags": [
          "ListingsService"
        ]
      },
      "post": {
        "summary": "Создание нового объявления",
        "description": "Создает новое объявление с указанным заголовком, текстом, изображением и ценой",
        "operationId": "ListingsService_CreateListing",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/listingsListingResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/listingsCreateListingRequest"
            }
          }
        ],
        "tags": [
          "ListingsService"
        ]
      }
    }
  },
  "definitions": {
    "listingsCreateListingRequest": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "imageUrl": {
          "type": "string"
        },
        "price": {
          "type": "number",
          "format": "float"
        }
      }
    },
    "listingsListingResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "uint64"
        },
        "title": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "imageUrl": {
          "type": "string"
        },
        "price": {
          "type": "number",
          "format": "float"
        },
        "authorUsername": {
          "type": "string"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "isOwner": {
          "type": "boolean"
        }
      }
    },
    "listingsListingsResponse": {
      "type": "object",
      "properties": {
        "listings": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/listingsListingResponse"
          }
        },
        "total": {
          "type": "integer",
          "format": "int64"
        },
        "page": {
          "type": "integer",
          "format": "int64"
        },
        "perPage": {
          "type": "integer",
          "format": "int64"
        },
        "totalPages": {
          "type": "integer",
          "format": "int64"
        }
      }
    },
    "listingsSortField": {
      "type": "string",
      "enum": [
        "SORT_FIELD_UNSPECIFIED",
        "SORT_FIELD_CREATED_AT",
        "SORT_FIELD_PRICE"
      ],
      "default": "SORT_FIELD_UNSPECIFIED"
    },
    "listingsSortOrder": {
      "type": "string",
      "enum": [
        "SORT_ORDER_UNSPECIFIED",
        "SORT_ORDER_ASC",
        "SORT_ORDER_DESC"
      ],
      "default": "SORT_ORDER_UNSPECIFIED"
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}